chore: extract showPinScreenForResult to helper.js
